webkit: Support video frame capture for HTML5 canvas use cases
Includes several patches cherry-picked from jb branch.

Implement MediaPlayerPrivate::hasSingleSecurityOrigin

Allows drawing video to the canvas as long as the video is not a
playlist file which may contain references from different origins. The
MediaPlayer should be the source of this knowledge.

TODO: Have the MediaPlayer check the data sources and report back to
WebKit if the playlist files have the same security origin.

Change-Id: I184c173ad454c02649d0a9930e20e037846bbe33

HTML5 video should still play when not visible

The WebGL 1.0.1 texture-npot-video test expects to be able to play video
when the video element has display:none. According to the HTML5 video
specification this is allowed. The video frames won't be painted but the
media player will still generate the frames and they can be read once
the drawImage and texImage2D APIs gain support for video.

Change-Id: Ia265a3bc8394c097ffa917809c592b6cf1ef86ec

Conflicts:

	Source/WebKit/android/WebCoreSupport/MediaPlayerPrivateAndroid.cpp

Allow HTMLMediaElement to capture video frames

Allow HTMLMediaElement to capture video frames by implementing
MediaPlayer::paintCurrentFrameInContext.  Because the MediaPlayer and
video texture are created in the UI context, need to do signalling
between the webcore and UI threads to do the video frame capture.

Modified android media player to return accurate ready states so
that webkit can check for valid states to issue the media frame capture.
In MediaPlayerPrivateAndroid, delay ready state change to "HaveEnoughData"
only when the first video frame is available.

Changed HTMLVideoElement to create a renderer even if style specifies
display "none" so that a corresponding platform layer can be created for
the RenderVideo. This allows canvas or WebGL to access video frames even
when the video itself is not displayed.

Change-Id: I396681d8f2ec7030a380ac4ffccbd4c95892eed0

Conflicts:

	Source/WebCore/platform/graphics/android/layers/VideoLayerAndroid.cpp
	Source/WebCore/platform/graphics/android/layers/VideoLayerAndroid.h
	Source/WebCore/platform/graphics/android/layers/VideoLayerManager.cpp
	Source/WebCore/platform/graphics/android/layers/VideoLayerManager.h
	Source/WebCore/platform/graphics/android/rendering/ShaderProgram.h
	Source/WebKit/android/WebCoreSupport/MediaPlayerPrivateAndroid.cpp

Add callback for player state changes from HTML5 video java proxy

Keeps VideoLayerAndroid's player state in sync with the HTML5 video java
proxy's media player state. This is required so that
paintCurrentFrameInContext calls do not capture empty video frames.

CRs-Fixed: 426323