Use combined depth and stencil buffers
Use combined depth and stencil buffers when both are requested together.
When stencil and depth buffers are requested, create a packed buffer
instead of standalone buffers and attach it to both the DEPTH_ATTACHMENT
point as well as the STENCIL_ATTACHMENT point as per
OES_packed_depth_stencil extension to GLES2.

CRs-fixed: 45920

Change-Id: I537efb611158935206c295fec8fc1a9b6d27e82b